DOC-1103 Refactored DocumentNameMapper
