+++ src/index.ts
@@ -1,11 +1,8 @@
 import { BlobServiceClient as BlobClient } from "@azure/storage-blob";
 import { S3Client } from "@aws-sdk/client-s3";
 import { config } from "./config";
-import {
-  BlobListStream,
-  BlobToS3CopyStream,
-  createS3ObjectNameMapper,
-} from "./storage";
+import { BlobListStream, BlobToS3CopyStream } from "./storage";
+import { createDocumentNameMapper } from "./document";
 import { ThrottledTransformStream } from "./throttle";
 
 const blobClient = new BlobClient(config.azure.blob.url);
@@ -22,10 +19,10 @@ const s3Client = new S3Client({
   },
   region: config.aws.region,
 });
-const s3ObjectNameMapper = createS3ObjectNameMapper({
-  blobPrefix: config.azure.blob.prefix,
-  s3Prefix: config.aws.s3.prefix,
-});
+const s3ObjectNameMapper = createDocumentNameMapper(
+  config.azure.blob.prefix,
+  config.aws.s3.prefix
+);
 const blobToS3CopyStream = new BlobToS3CopyStream({
   blobClient,
   blobContainerName: config.azure.blob.containerName,
